% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\name{counter_get_value}
\alias{counter_get_value}
\title{Get the value from a trait-wrapped counter}
\source{
Generated by miniextendr from Rust fn \code{counter_get_value}
}
\usage{
counter_get_value(counter_sexp)
}
\arguments{
\item{counter_sexp}{An external pointer from new_counter()}
}
\value{
The counter's current value
}
\description{
Get the value from a trait-wrapped counter
}
