% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{debug_shared_data_type_name}
\alias{debug_shared_data_type_name}
\title{Debug: Get the ExternalPtr type name for SharedData}
\source{
Generated by miniextendr from Rust fn \code{debug_shared_data_type_name}
}
\usage{
debug_shared_data_type_name()
}
\description{
Debug: Get the ExternalPtr type name for SharedData
}
