% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{debug_tag_counter}
\alias{debug_tag_counter}
\title{Debug: Get TAG_COUNTER as hex string}
\source{
Generated by miniextendr from Rust fn \code{debug_tag_counter}
}
\usage{
debug_tag_counter()
}
\description{
Debug: Get TAG_COUNTER as hex string
}
