% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{EnvPoint}
\alias{EnvPoint}
\alias{EnvPoint$new}
\alias{EnvPoint$x}
\alias{EnvPoint$y}
\alias{EnvPoint$distance_from_origin}
\alias{EnvPoint$add}
\alias{$.EnvPoint}
\alias{[[.EnvPoint}
\title{EnvPoint (Env-style)}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{EnvPoint}
}
\usage{
EnvPoint

\method{$}{EnvPoint}(self, name)

\method{[[}{EnvPoint}(self, name)
}
\description{
A 2D point using the default Env-style class system.

Create a new point

Get x coordinate

Get y coordinate

Calculate distance from origin

Add another point's coordinates
}
\examples{
p <- EnvPoint$new(3.0, 4.0)
p$x()
p$y()
p$distance_from_origin()
}
\keyword{datasets}
