% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-package.R
\docType{package}
\name{producer.pkg-package}
\alias{producer.pkg}
\alias{producer.pkg-package}
\title{producer.pkg: Cross-Package Trait Dispatch Producer Example}
\description{
Demonstrates miniextendr's cross-package trait dispatch capabilities. Defines the Counter trait and implements it for SimpleCounter. Objects created here can be passed to consumer.pkg functions that operate on the Counter trait interface, proving ABI compatibility via type tag matching.
}
\author{
\strong{Maintainer}: Mossa M. Reimert \email{mossa@sund.ku.dk}

Other contributors:
\itemize{
  \item A2-AI [copyright holder, funder]
}

}
\keyword{internal}
