% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{s3point_add}
\alias{s3point_add}
\title{S3 generic for \code{s3point_add}}
\source{
Generated by miniextendr from \code{S3Point::s3point_add}
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
S3 generic for \code{s3point_add}
}
