% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\name{s4_distance}
\alias{s4_distance}
\title{The \code{s4_distance()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s4_distance()} is an S4 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S4Point}}
}
}
