% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\docType{class}
\name{S4Point}
\alias{S4Point}
\alias{s4point_x}
\alias{s4point_y}
\alias{s4point_distance}
\alias{S4Point-s4_x}
\alias{s4_x,S4Point-method}
\alias{S4Point-s4_y}
\alias{s4_y,S4Point-method}
\alias{S4Point-s4_distance}
\alias{s4_distance,S4Point-method}
\alias{S4Point-s4_add}
\alias{s4_add,S4Point-method}
\title{S4Point (S4-style)}
\source{
Generated by miniextendr from Rust type \code{S4Point}

Generated by miniextendr from \code{S4Point::new}

Generated by miniextendr from \code{S4Point::s4_x}

Generated by miniextendr from \code{S4Point::s4_y}

Generated by miniextendr from \code{S4Point::s4_distance}

Generated by miniextendr from \code{S4Point::s4_add}
}
\usage{
S4Point(x, y)
}
\arguments{
\item{x}{(undocumented)}

\item{y}{(undocumented)}
}
\description{
A 2D point using S4-style class system.

Create a new point

Get x coordinate

Get y coordinate

Calculate distance from origin

Add to coordinates
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4Point} struct}
}}

