% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\name{s7point_x}
\alias{s7point_x}
\title{The \code{s7point_x()} generic}
\arguments{
\item{x}{An object.}

\item{...}{Passed on to methods.}
}
\description{
\code{s7point_x()} is an S7 generic generated by miniextendr. Methods in this
package are available for:
\itemize{
\item \code{\link{S7Point}}
}
}
