% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SharedData}
\alias{SharedData}
\alias{SharedData$get_x}
\alias{SharedData$get_y}
\alias{SharedData$get_label}
\alias{SharedData$create}
\alias{$.SharedData}
\alias{[[.SharedData}
\title{SharedData (Env-style)}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{SharedData}
}
\usage{
SharedData

\method{$}{SharedData}(self, name)

\method{[[}{SharedData}(self, name)
}
\description{
A point with label using Env-style class system.

Get the x coordinate

Get the y coordinate

Get the label

Create a new SharedData
}
\examples{
p <- SharedData$create(1.0, 2.0, "origin")
p$get_x()
p$get_y()
p$get_label()
}
\keyword{datasets}
