% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\name{StatefulCounter}
\alias{StatefulCounter}
\alias{StatefulCounter$get_value}
\alias{StatefulCounter$history_len}
\alias{$.StatefulCounter}
\alias{[[.StatefulCounter}
\alias{StatefulCounter$Counter$value}
\alias{StatefulCounter$Counter$increment}
\alias{StatefulCounter$Counter$add}
\alias{StatefulCounter$Counter$panic_plain}
\alias{StatefulCounter$Counter$error_with_class}
\alias{StatefulCounter$Counter$raise_error}
\alias{StatefulCounter$Counter$raise_warning}
\alias{StatefulCounter$Counter$raise_warning_without_call}
\alias{StatefulCounter$Counter$raise_message}
\alias{StatefulCounter$Counter$raise_condition_classed}
\alias{StatefulCounter$Counter$raise_error_with_data}
\alias{StatefulCounter$Counter$raise_arg_error}
\alias{StatefulCounter$Resettable$reset}
\alias{StatefulCounter$Resettable$is_default}
\title{StatefulCounter Class}
\source{
Generated by miniextendr from Rust type \code{StatefulCounter}

Generated by miniextendr from \code{StatefulCounter::get_value}

Generated by miniextendr from \code{StatefulCounter::history_len}
}
\usage{
StatefulCounter

\method{$}{StatefulCounter}(self, name)

\method{[[}{StatefulCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get current value (inherent method)

Get how many history entries exist
}
