% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\docType{data}
\name{EnvPoint}
\alias{EnvPoint}
\alias{EnvPoint$new}
\alias{EnvPoint$x}
\alias{EnvPoint$y}
\alias{EnvPoint$distance_from_origin}
\alias{EnvPoint$add}
\alias{$.EnvPoint}
\alias{[[.EnvPoint}
\title{EnvPoint (Env-style)}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{EnvPoint}

Generated by miniextendr from \code{EnvPoint::new}

Generated by miniextendr from \code{EnvPoint::x}

Generated by miniextendr from \code{EnvPoint::y}

Generated by miniextendr from \code{EnvPoint::distance_from_origin}

Generated by miniextendr from \code{EnvPoint::add}
}
\usage{
EnvPoint

\method{$}{EnvPoint}(self, name)

\method{[[}{EnvPoint}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
A 2D point using the default Env-style class system.

Create a new point

Get x coordinate

Get y coordinate

Calculate distance from origin

Add another point's coordinates
}
\examples{
p <- EnvPoint$new(3.0, 4.0)
p$x()
p$y()
p$distance_from_origin()
}
\keyword{datasets}
