% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\docType{data}
\name{S3Point}
\alias{S3Point}
\alias{new_s3point}
\alias{s3point_x}
\alias{s3point_y}
\alias{s3point_distance}
\alias{s3point_x.S3Point}
\alias{s3point_y.S3Point}
\alias{s3point_distance.S3Point}
\alias{s3point_add.S3Point}
\title{S3Point (S3-style)}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from Rust type \code{S3Point}

Generated by miniextendr from \code{S3Point::s3point_x}

Generated by miniextendr from \code{S3Point::s3point_x}

Generated by miniextendr from \code{S3Point::s3point_y}

Generated by miniextendr from \code{S3Point::s3point_y}

Generated by miniextendr from \code{S3Point::s3point_distance}

Generated by miniextendr from \code{S3Point::s3point_distance}

Generated by miniextendr from \code{S3Point::s3point_add}

Generated by miniextendr from \code{S3Point::s3point_add}
}
\usage{
new_s3point(x, y)

\method{s3point_x}{S3Point}(x, ...)

\method{s3point_y}{S3Point}(x, ...)

\method{s3point_distance}{S3Point}(x, ...)

\method{s3point_add}{S3Point}(x, dx, dy, ...)

S3Point
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments.}

\item{dx}{(undocumented)}

\item{dy}{(undocumented)}
}
\description{
A 2D point using S3-style class system.

Create a new point

Get x coordinate

Get y coordinate

Calculate distance from origin

Add to coordinates
}
\keyword{datasets}
