% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{class}
\name{S4Point}
\alias{S4Point}
\alias{s4point_x}
\alias{s4point_y}
\alias{s4point_distance}
\alias{s4_x}
\alias{s4_y}
\alias{s4_distance}
\alias{s4_add}
\title{S4Point (S4-style)}
\source{
Generated by miniextendr from Rust type \code{S4Point}

Generated by miniextendr from \code{S4Point::new}

Generated by miniextendr from \code{S4Point::x}

Generated by miniextendr from \code{S4Point::y}

Generated by miniextendr from \code{S4Point::distance}

Generated by miniextendr from \code{S4Point::add}
}
\usage{
S4Point(x, y)
}
\description{
A 2D point using S4-style class system.

Create a new point

Get x coordinate

Get y coordinate

Calculate distance from origin

Add to coordinates
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to Rust \code{S4Point} struct}
}}

\examples{
p <- S4Point(3.0, 4.0)
s4point_x(p)
s4point_y(p)
s4point_distance(p)
}
