% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer.pkg-wrappers.R
\name{SimpleCounter}
\alias{SimpleCounter}
\alias{SimpleCounter$get_value}
\alias{$.SimpleCounter}
\alias{[[.SimpleCounter}
\alias{SimpleCounter$Counter$value}
\alias{SimpleCounter$Counter$increment}
\alias{SimpleCounter$Counter$add}
\alias{SimpleCounter$Counter$panic_plain}
\alias{SimpleCounter$Counter$error_with_class}
\alias{SimpleCounter$Counter$raise_error}
\alias{SimpleCounter$Counter$raise_warning}
\alias{SimpleCounter$Counter$raise_message}
\alias{SimpleCounter$Counter$raise_condition_classed}
\alias{SimpleCounter$Counter$raise_error_with_data}
\alias{SimpleCounter$Resettable$reset}
\alias{SimpleCounter$Resettable$is_default}
\title{SimpleCounter  Class}
\source{
Generated by miniextendr from Rust type \code{SimpleCounter}

Generated by miniextendr from \code{SimpleCounter::get_value}
}
\usage{
SimpleCounter

\method{$}{SimpleCounter}(self, name)

\method{[[}{SimpleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get current value (inherent method)
}
