% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{SimpleCounter}
\alias{SimpleCounter}
\alias{SimpleCounter$get_value}
\alias{$.SimpleCounter}
\alias{[[.SimpleCounter}
\alias{SimpleCounter$Counter$value}
\alias{SimpleCounter$Counter$increment}
\alias{SimpleCounter$Counter$add}
\title{SimpleCounter  Class}
\format{
An object of class \code{environment} of length 2.
}
\source{
Generated by miniextendr from Rust type \code{SimpleCounter}

Generated by miniextendr from \code{SimpleCounter::get_value}
}
\usage{
SimpleCounter

\method{$}{SimpleCounter}(self, name)

\method{[[}{SimpleCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get current value (inherent method)
}
\keyword{datasets}
