% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producer-pkg-wrappers.R
\docType{data}
\name{StatefulCounter}
\alias{StatefulCounter}
\alias{StatefulCounter$get_value}
\alias{StatefulCounter$history_len}
\alias{$.StatefulCounter}
\alias{[[.StatefulCounter}
\alias{StatefulCounter$Counter$value}
\alias{StatefulCounter$Counter$increment}
\alias{StatefulCounter$Counter$add}
\alias{StatefulCounter$Resettable$reset}
\alias{StatefulCounter$Resettable$is_default}
\title{StatefulCounter  Class}
\format{
An object of class \code{environment} of length 4.
}
\source{
Generated by miniextendr from Rust type \code{StatefulCounter}

Generated by miniextendr from \code{StatefulCounter::get_value}

Generated by miniextendr from \code{StatefulCounter::history_len}
}
\usage{
StatefulCounter

\method{$}{StatefulCounter}(self, name)

\method{[[}{StatefulCounter}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Get current value (inherent method)

Get how many history entries exist
}
\keyword{datasets}
